+++ changelog.md
@@ -9,6 +9,7 @@ Beta
 - Perfect Parrying attacks to deal counter damage
 - Moved all UP/DOWN vector constants to ShapeUtil
 - Added "/titan stat" sub-command
+- Added Titan footstep sounds
 - Tweaked sprint charging so it only uses stamina if you've shed enough armor
 - Added a few sound effects for titan combat and odm blade swinging
 
@@ -16,3 +17,4 @@ TODO:
 - Investigate why blade handles stop swinging randomly
 - Titan endurance (maybe? still pondering)
 - ODM hooks not attaching to entities properly
+- Grabbing